@@ -3,18 +3,20 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
-using Newtonsoft.Json;
+using Navigator.Configuration.Options;
+using Navigator.Strategy;
+using Telegram.Bot;
 using Telegram.Bot.Types;
 
 namespace Navigator.Configuration;
 
 /// <summary>
-/// Navigator extensions for <see cref="IEndpointRouteBuilder"/>.
+///     Navigator extensions for <see cref="IEndpointRouteBuilder" />.
 /// </summary>
 public static class EndpointRouteBuilderExtensions
 {
     /// <summary>
-    /// Configure navigator's provider's endpoints.
+    ///     Configure navigator's provider's endpoints.
     /// </summary>
     /// <param name="endpointRouteBuilder"></param>
     /// <returns></returns>
@@ -26,29 +28,27 @@ public static class EndpointRouteBuilderExtensions
 
         endpointRouteBuilder.MapPost(options.GetWebHookEndpointOrDefault(), ProcessTelegramUpdate);
     }
-        
+
     private static async Task ProcessTelegramUpdate(HttpContext context)
     {
         context.Response.StatusCode = 200;
 
-        if (context.Request.ContentType != "application/json")
-        {
-            return;
-        }
+        if (context.Request.ContentType != "application/json") return;
 
         var telegramUpdate = await ParseTelegramUpdate(context.Request);
 
-        var navigatorMiddleware = context.RequestServices.GetRequiredService<TelegramMiddleware>();
+        var strategy = context.RequestServices.GetRequiredService<INavigatorStrategy>();
 
-        await navigatorMiddleware.Process(telegramUpdate);
+        await strategy.Invoke(telegramUpdate);
     }
-        
+
     private static async Task<Update> ParseTelegramUpdate(HttpRequest request)
     {
         try
         {
             var reader = new StreamReader(request.Body);
-            return JsonConvert.DeserializeObject<Update>(await reader.ReadToEndAsync()) ?? throw new InvalidOperationException();
+            return JsonSerializer.Deserialize<Update>(await reader.ReadToEndAsync(), JsonBotAPI.Options) ??
+                   throw new InvalidOperationException();
         }
         catch (Exception e)
         {
