@@ -1,23 +1,23 @@
-using Microsoft.AspNetCore.Routing;
-
-namespace Navigator.Configuration;
-
-/// <summary>
-/// Configures the route for the telegram endpoint.
-/// </summary>
-public class NavigatorRouteConfiguration
-{
-    /// <summary>
-    /// Endpoint Route Builder.
-    /// </summary>
-    public IEndpointRouteBuilder EndpointRouteBuilder { get; internal set; }
-
-    /// <summary>
-    /// Default constructor.
-    /// </summary>
-    /// <param name="endpointRouteBuilder"></param>
-    public NavigatorRouteConfiguration(IEndpointRouteBuilder endpointRouteBuilder)
-    {
-        EndpointRouteBuilder = endpointRouteBuilder;
-    }
-}
+// using Microsoft.AspNetCore.Routing;
+//
+// namespace Navigator.Configuration;
+//
+// /// <summary>
+// /// Configures the route for the telegram endpoint.
+// /// </summary>
+// public class NavigatorRouteConfiguration
+// {
+//     /// <summary>
+//     /// Endpoint Route Builder.
+//     /// </summary>
+//     public IEndpointRouteBuilder EndpointRouteBuilder { get; internal set; }
+//
+//     /// <summary>
+//     /// Default constructor.
+//     /// </summary>
+//     /// <param name="endpointRouteBuilder"></param>
+//     public NavigatorRouteConfiguration(IEndpointRouteBuilder endpointRouteBuilder)
+//     {
+//         EndpointRouteBuilder = endpointRouteBuilder;
+//     }
+// }
