@@ -1,25 +1,22 @@
-using MediatR;
 using Microsoft.Extensions.DependencyInjection;
-using Navigator.Actions;
-using Navigator.Bundled.Extensions.Update;
+using Navigator.Catalog.Factory;
 using Navigator.Client;
 using Navigator.Configuration;
-using Navigator.Context;
-using Navigator.Context.Accessor;
-using Navigator.Context.Builder;
-using Navigator.Extensions;
+using Navigator.Configuration.Options;
 using Navigator.Hosted;
-using Scrutor;
+using Navigator.Strategy;
+using Navigator.Strategy.Classifier;
+using Navigator.Strategy.TypeProvider;
 
 namespace Navigator;
 
 /// <summary>
-/// Extensions for configuring Navigator.
+///     Extensions for configuring Navigator.
 /// </summary>
 public static class ServiceCollectionExtensions
 {
     /// <summary>
-    /// Adds Navigator.
+    ///     Adds Navigator.
     /// </summary>
     /// <param name="services"></param>
     /// <param name="options"></param>
@@ -28,50 +25,27 @@ public static class ServiceCollectionExtensions
     public static NavigatorConfiguration AddNavigator(this IServiceCollection services, Action<NavigatorOptions> options)
     {
         if (options == null)
-        {
             throw new ArgumentNullException(nameof(options), "Navigator options are required for navigator framework to work.");
-        }
 
         var navigatorBuilder = new NavigatorConfiguration(options, services);
 
-        services.AddNavigatorContextServices();
-        
-        services.AddScoped<TelegramMiddleware>();
-
         services.AddScoped<INavigatorClient, NavigatorClient>();
-        
-        services.AddScoped<INavigatorContextExtension, UpdateNavigatorContextExtension>();
 
-        services.AddScoped<IActionLauncher, ActionLauncher>();
+        services.AddSingleton<BotActionCatalogFactory>();
+
+        services.AddScoped<IUpdateClassifier, UpdateClassifier>();
+
+        services.AddScoped<IArgumentTypeProvider, NavigatorEntitiesTypeProvider>();
+        services.AddScoped<IArgumentTypeProvider, TelegramEntitiesTypeProvider>();
+        services.AddScoped<IArgumentTypeProvider, TelegramMessageTypeProvider>();
+        services.AddScoped<IArgumentTypeProvider, TelegramUpdateTypeProvider>();
+
+        services.AddScoped<INavigatorStrategy, NavigatorStrategy>();
 
         services.AddHostedService<SetTelegramBotWebHookHostedService>();
-
-        services.AddMediatR(navigatorBuilder.Options.GetActionsAssemblies());
-
-        services.Scan(scan => scan
-            .FromAssemblies(navigatorBuilder.Options.GetActionsAssemblies())
-            .AddClasses(classes => classes.AssignableTo<IAction>())
-            .UsingRegistrationStrategy(RegistrationStrategy.Append)
-            .AsSelf()
-            .WithScopedLifetime());
-
-        navigatorBuilder.Options.RegisterActionsCore(services
-            .Where(descriptor => descriptor.ImplementationType?.IsAssignableTo(typeof(IAction)) ?? false)
-            .Select(descriptor => descriptor.ImplementationType!));
-        
-        navigatorBuilder.Options.RegisterActionsPriorityCore(services
-            .Where(descriptor => descriptor.ImplementationType?.IsAssignableTo(typeof(IAction)) ?? false)
-            .Select(descriptor => descriptor.ImplementationType!));
 
         navigatorBuilder.RegisterOrReplaceOptions();
 
         return navigatorBuilder;
-    }
-
-    internal static void AddNavigatorContextServices(this IServiceCollection services)
-    {
-        services.AddScoped<INavigatorContextBuilder, NavigatorContextBuilder>();
-        services.AddScoped<INavigatorContextFactory, NavigatorContextFactory>();
-        services.AddTransient<INavigatorContextAccessor, NavigatorContextAccessor>();
     }
 }
